Dedup container exception
